Atualizando regras de status para solicitacoes e budgets

+++ src/domain/portal/application/use-cases/solicitations/delete.solicitation-use-case.ts
@@ -3,6 +3,8 @@ import { UnauthorizedSolicitationActionError } from "../../errors/repair/solicit
 import { SolicitationNotFoundError } from "../../errors/repair/solicitations/SolicitationNotFoundError"
 import { ISolicitationRepository } from "../../repositories/repair/solicitation-repository.interface"
 import { ISolicitationFormRepository } from "../../repositories/repair/solicitation-form.repository.interface"
+import { IBudgetRepository } from "../../repositories/repair/budget-repository"
+import { ActionNotAllowedError } from "../../errors/repair/solicitations/ActionNotAllowed"
 
 type DeleteSolicitationUseCaseResponse = Either<
   SolicitationNotFoundError | UnauthorizedSolicitationActionError,
@@ -17,6 +19,7 @@ export class DeleteSolicitationUseCase {
   constructor(
     private solicitationRepository: ISolicitationRepository,
     private solicitationFormRepository: ISolicitationFormRepository,
+    private budgetRepository: IBudgetRepository,
   ) {}
 
   async execute({
@@ -26,7 +29,10 @@ export class DeleteSolicitationUseCase {
     const solicitation =
       await this.solicitationRepository.findById(solicitationId)
     if (!solicitation) return left(new SolicitationNotFoundError())
-
+    const budget = await this.budgetRepository.findByParam<{
+      solicitationId: string
+    }>({ solicitationId: solicitation.id.toString() })
+    if (budget.length > 0) return left(new ActionNotAllowedError())
     if (solicitation.clientProfile.id.toString() !== profileId)
       return left(new UnauthorizedSolicitationActionError())
     await this.solicitationRepository.deleteById(solicitationId)
